fix(extension): better path handling
Hopefully that fixes the windows build.
