+++ crates/extension/Cargo.toml
@@ -15,9 +15,4 @@ semver = { workspace = true, features = ["serde"] }
 grafbase-workspace-hack.workspace = true
 serde.workspace = true
 url.workspace = true
-reqwest.workspace = true
-serde_json.workspace = true
 
-[dev-dependencies]
-tokio.workspace = true
-tempfile = "3"
